requirements.txt
# Core dependencies
aiogram==3.7.0
requests>=2.31.0
python-dotenv>=1.0.0

# Payment providers (optional, install as needed)
# YooKassa
yookassa>=3.0.0

# CryptoPay
aiocryptopay>=0.4.0
